requirements.txt
django==1.11.7
psycopg2==2.7.3.2
django-import-export==0.6.1